test: Refactor to move tests from Transaction to TransactionRow
