+++ src/ducks/transactions/Transactions.spec.jsx
@@ -1,87 +1,15 @@
 /* global mount */
 
 import React from 'react'
-import { render } from '@testing-library/react'
-import TransactionRowDesktop from './TransactionRow/TransactionRowDesktop'
-import TransactionRowMobile from './TransactionRow/TransactionRowMobile'
-import { TransactionsDumb, sortByDate } from './Transactions'
-import data from '../../../test/fixtures'
-import store from '../../../test/store'
-import AppLike from '../../../test/AppLike'
-import { getClient } from 'test/client'
-import { normalizeData } from 'test/store'
+
+import data from 'test/fixtures'
+import AppLike from 'test/AppLike'
+
 import TransactionPageErrors from 'ducks/transactions/TransactionPageErrors'
+import { TransactionsDumb, sortByDate } from './Transactions'
 
 // No need to test this here
 jest.mock('ducks/transactions/TransactionPageErrors', () => () => null)
-jest.mock('react-tappable/lib/Tappable', () => ({ children }) => children)
-
-const allTransactions = data['io.cozy.bank.operations']
-
-describe('transaction row', () => {
-  let client, transaction
-
-  const setup = (row, withTable) => {
-    return render(
-      <AppLike store={store} client={client}>
-        {withTable ? (
-          <table>
-            <tbody>{row}</tbody>
-          </table>
-        ) : (
-          row
-        )}
-      </AppLike>
-    )
-  }
-
-  const rawTransaction = allTransactions[0]
-  rawTransaction._type = 'io.cozy.bank.operations'
-
-  beforeEach(() => {
-    client = getClient()
-    client.ensureStore()
-    const datastore = normalizeData({
-      'io.cozy.bank.accounts': data['io.cozy.bank.accounts']
-    })
-    jest
-      .spyOn(client, 'getDocumentFromState')
-      .mockImplementation((doctype, id) => {
-        return datastore[doctype][id]
-      })
-    transaction = client.hydrateDocument(rawTransaction)
-  })
-
-  it('should render correctly on desktop', () => {
-    const handleRef = jest.fn()
-    const root = setup(
-      <TransactionRowDesktop
-        transaction={transaction}
-        urls={{}}
-        brands={[]}
-        onRef={handleRef}
-      />,
-      true
-    )
-    expect(root.getByText('Compte courant Isabelle - BNPP')).toBeTruthy()
-  })
-
-  it('should render correctly on mobile', () => {
-    const handleRef = jest.fn()
-    const root = setup(
-      <TransactionRowMobile
-        onRef={handleRef}
-        transaction={transaction}
-        urls={{}}
-        brands={[]}
-      />,
-      false
-    )
-
-    expect(root.getByText('Compte courant Isabelle - BNPP')).toBeTruthy()
-    expect(handleRef).toHaveBeenCalled()
-  })
-})
 
 describe('Transactions', () => {
   let i = 0
